+++ frontend/src/pages/ReviewWritingCardFromPage/components/CardSliderController/index.tsx
