+++ frontend/src/pages/ReviewWritingCardFromPage/components/QnABox/style.ts
