+++ frontend/src/pages/ReviewWritingCardFromPage/index.tsx
@@ -1,7 +1,7 @@
 import { ErrorSuspenseContainer } from '@/components';
 
-import CardForm from './CardForm';
-// TODO : API 연결 후 ReviewWriting 폴더 삭제
+import CardForm from './components/CardForm';
+
 const ReviewWritingCardFormPage = () => {
   return (
     <ErrorSuspenseContainer>
